+++ src/test/util_test.ts
@@ -122,6 +122,36 @@ suite('MergedAsyncIterables', () => {
     const actual = await flush(merged);
     assert.deepEqual(actual, expected);
   });
+
+  test('notice a new iterator while waiting on an existing one', async () => {
+    const merged = new MergedAsyncIterables<string>();
+
+    merged.add(
+      (async function* () {
+        await raf();
+        yield 'slow';
+      })()
+    );
+
+    // The key thing about this test is that we've started iterating before
+    // adding the second source.
+    const actual: string[] = [];
+    const done = (async () => {
+      for await (const value of merged) {
+        actual.push(value);
+      }
+    })();
+
+    merged.add(
+      (async function* () {
+        yield 'fast';
+      })()
+    );
+
+    await done;
+    const expected = ['fast', 'slow'];
+    assert.deepEqual(actual, expected);
+  });
 });
 
 suite('relativeUrlPath', () => {