+++ src/playground-ide.ts
@@ -13,7 +13,7 @@ import './playground-file-editor.js';
 import './playground-preview.js';
 import {PlaygroundProject} from './playground-project.js';
 import {ProjectManifest} from './shared/worker-api.js';
-import {version} from './shared/version.js';
+import {npmVersion} from './shared/version.js';
 
 /**
  * A multi-file code editor component with live preview that works without a
@@ -209,7 +209,7 @@ export class PlaygroundIde extends LitElement {
    * "/node_modules/playground-elements/").
    */
   @property({attribute: 'sandbox-base-url'})
-  sandboxBaseUrl = `https://unpkg.com/playground-elements@${version}/`;
+  sandboxBaseUrl = `https://unpkg.com/playground-elements@${npmVersion}/`;
 
   /**
    * The service worker scope to register on